requirements.txt
python-dotenv
groq
streamlit
requests
pandas
plotly
requests